server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -6,9 +6,10 @@ const NotFoundError = require('../../exceptions/NotFoundError');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class NotesService {
-  constructor(collaborationService) {
+  constructor(collaborationService, chacheService) {
     this._pool = new Pool();
     this._collaborationService = collaborationService;
+    this._chacheService = chacheService;
   }
 
   async addNote({ title, body, tags, owner }) {
@@ -27,17 +28,32 @@ class NotesService {
       throw new InvariantError('Catatan gagal ditambahkan');
     }
 
+    await this._chacheService.del(`notes:${owner}`);
+
     return result.rows[0].id;
   }
 
   async getNotes(owner) {
-    const query = {
-      text: 'SELECT notes.* FROM notes LEFT JOIN collaborations ON collaborations.note_id = notes.id WHERE notes.owner = $1 OR collaborations.user_id = $1 GROUP BY notes.id',
-      values: [owner],
-    };
-    const result = await this._pool.query(query);
+    try {
+      const result = await this._chacheService.get(`notes:${owner}`);
 
-    return result.rows.map(mapDBToModel);
+      return JSON.parse(result);
+    } catch {
+      const query = {
+        text: 'SELECT notes.* FROM notes LEFT JOIN collaborations ON collaborations.note_id = notes.id WHERE notes.owner = $1 OR collaborations.user_id = $1 GROUP BY notes.id',
+        values: [owner],
+      };
+
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModel);
+
+      await this._chacheService.set(
+        `notes:${owner}`,
+        JSON.stringify(mappedResult)
+      );
+
+      return mappedResult;
+    }
   }
 
   async getNoteById(id) {
@@ -70,7 +86,7 @@ class NotesService {
     const updateAt = new Date().toISOString();
 
     const query = {
-      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, update_at = $4 WHERE id = $5 RETURNING id',
+      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, update_at = $4 WHERE id = $5 RETURNING id, owner',
       values: [title, body, tags, updateAt, id],
     };
 
@@ -79,11 +95,15 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan');
     }
+
+    const { owner } = result.rows[0];
+
+    await this._chacheService.del(`notes:${owner}`);
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
@@ -92,6 +112,11 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan');
     }
+
+    const { owner } = result.rows[0];
+
+    await this._chacheService.del(`notes:${owner}`);
+
   }
 
   async verifyNoteOwner(id, owner) {
